rename users.js to login.js
